# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Darwin Ops MCP Backend API with Tool Registry, Policy Engine, and Approval Flow.\nReturns immediate execution result or 202 + approval ID for high-risk tools.",
        "title": "Darwin Ops MCP API",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080",
            "email": "support@darwin-ops-mcp.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/agent-keys": {
            "get": {
                "description": "Lists API key metadata. Plaintext secrets and hashes are never returned.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-keys"
                ],
                "summary": "List Agent API Keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Issues a bearer token for an AI agent. Requires the master API token when authentication is enabled.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-keys"
                ],
                "summary": "Create Agent API Key",
                "parameters": [
                    {
                        "description": "Agent API key issuance request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agent-keys/{id}/revoke": {
            "post": {
                "description": "Revokes an agent API key by ID. Revoked keys can no longer authenticate.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-keys"
                ],
                "summary": "Revoke Agent API Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/applications": {
            "get": {
                "description": "Returns all tool access applications in creation order (newest last).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "List Applications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Applies for access to a tool with a specific risk level, role, and reason.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Submit Tool Application",
                "parameters": [
                    {
                        "description": "Tool application request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/applications/{id}/approve": {
            "post": {
                "description": "Approves a pending tool access application and records the decision timestamp.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Approve Tool Application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.ToolApplication"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/applications/{id}/reject": {
            "post": {
                "description": "Rejects a pending tool access application and records the decision timestamp.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Reject Tool Application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.ToolApplication"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/approvals/{id}/approve": {
            "post": {
                "description": "Marks a pending execution approval as approved.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Approve Execution Approval",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Approval ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Approval"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/approvals/{id}/reject": {
            "post": {
                "description": "Marks a pending execution approval as rejected.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Reject Execution Approval",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Approval ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Approval"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/executions/{id}": {
            "get": {
                "description": "Returns an execution record by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "executions"
                ],
                "summary": "Get Execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Execution"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tools": {
            "get": {
                "description": "Returns all registered tools.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "List Tools",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Tool"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a custom tool definition in the runtime registry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Create Tool",
                "parameters": [
                    {
                        "description": "Tool definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tools/{name}": {
            "get": {
                "description": "Returns a single registered tool by name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Get Tool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tool name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Tool"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing tool definition in the runtime registry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Update Tool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tool name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tool definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an existing tool definition from the runtime registry",
                "tags": [
                    "tools"
                ],
                "summary": "Delete Tool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tool name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tools/{name}/execute": {
            "post": {
                "description": "Executes a tool with the provided parameters. Returns 202 if approval",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Execute Tool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tool name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Execution request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Approval": {
            "type": "object",
            "properties": {
                "actor": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "decidedAt": {
                    "type": "string"
                },
                "executionId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.ApprovalStatus"
                },
                "target": {
                    "type": "string"
                },
                "tool": {
                    "type": "string"
                }
            }
        },
...
        "github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Execution": {
            "type": "object",
            "properties": {
                "actor": {
                    "type": "string"
                },
                "auditId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "reason": {
                    "type": "string"
                },
                "result": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "role": {
                    "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Role"
                },
                "status": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                },
                "tool": {
                    "type": "string"
                }
            }
        },
...
        "github_com_zlylong_darwin-ops-mcp_backend_internal_domain.RiskLevel": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high",
                "critical"
            ],
            "x-enum-varnames": [
                "RiskLow",
                "RiskMedium",
                "RiskHigh",
                "RiskCritical"
            ]
        },
        "github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Role": {
            "type": "string",
            "enum": [
                "viewer",
                "operator",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleViewer",
                "RoleOperator",
                "RoleAdmin"
            ]
        },
        "github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Tool": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "inputSchema": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.ParamSchema"
                    }
                },
                "name": {
                    "type": "string"
                },
                "readOnly": {
                    "type": "boolean"
                },
                "requiresApproval": {
                    "type": "boolean"
                },
                "risk": {
                    "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.RiskLevel"
                }
            }
        },
        "github_com_zlylong_darwin-ops-mcp_backend_internal_domain.ToolApplication": {
            "type": "object",
            "properties": {
                "actor": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "decidedAt": {
                    "type": "string"
                },
                "decision": {
                    "type": "string"
                },
                "durationHrs": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "reason": {
                    "type": "string"
                },
                "risk": {
                    "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.RiskLevel"
                },
                "role": {
                    "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.Role"
                },
                "status": {
                    "$ref": "#/definitions/github_com_zlylong_darwin-ops-mcp_backend_internal_domain.ApplicationStatus"
                },
                "tool": {
                    "type": "string"
                }
            }
        }
    }
}